requirements.txt
absl-py==2.1.0
accelerate==0.29.3
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.2
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asttokens==2.4.1
attrdict==2.0.1
attrs==23.2.0
av==13.0.0
binpacking==1.5.2
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
cmake==3.30.5
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cpm-kernels==1.0.11
crcmod==1.7
cryptography==42.0.5
cycler==0.12.1
dacite==1.8.1
dashscope==1.17.1
datasets==2.18.0
debugpy==1.8.9
decorator==5.1.1
deepseek_vl==1.0.0
deepspeed==0.14.4
dill==0.3.8
diskcache==5.6.3
docker-pycreds==0.4.0
docstring_parser==0.16
editdistance==0.8.1
einops==0.7.0
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.110.2
ffmpy==0.3.2
filelock==3.13.4
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
gradio==4.27.0
gradio_client==0.15.1
grpcio==1.62.2
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.24.6
humanfriendly==10.0
icecream==2.1.3
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
interegular==0.3.3
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.4.0
jsonlines==4.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
lark==1.2.2
llmuses==0.3.0
llvmlite==0.43.0
lm-format-enforcer==0.9.8
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
modelscope==1.15.0
mpmath==1.3.0
ms-swift==2.1.1.post2
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai==0.28.0
opencv-python==4.10.0.84
openpyxl==3.1.5
optimum==1.19.0
orjson==3.10.1
oss2==2.18.4
outlines==0.0.34
packaging==24.0
pandas==2.2.2
parso==0.8.4
peft==0.11.1
pexpect==4.9.0
pillow==10.3.0
pip==23.3.1
platformdirs==4.2.0
plotly==5.21.0
ply==3.11
portalocker==2.8.2
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==16.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
Pympler==1.0.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.2.0
qwen-vl-utils==0.0.8
ray==2.37.0
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rouge==1.0.1
rouge-chinese==1.0.3
rouge_score==0.1.2
rpds-py==0.18.0
ruff==0.4.1
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
sentry-sdk==1.45.0
setproctitle==1.3.3
setuptools==68.2.2
shellingham==1.5.4
shtab==1.7.1
simple-ddl-parser==1.1.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.37.2
streamlit==1.33.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.4.0
tiktoken==0.6.0
timm==1.0.7
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.39.3
transformers-stream-generator==0.0.5
triton==2.2.0
trl==0.8.5
typer==0.12.3
typing_extensions==4.11.0
tyro==0.8.3
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.21.0
wandb==0.16.6
watchdog==4.0.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.2
wheel==0.41.2
widgetsnbextension==4.0.13
xformers==0.0.25
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
